Fix missing non-null marker in 'Input: Structure' section
This seems to be a missing, presumably due to a typo, since removing with a nullable input seems incorrect. This PR adds the explanation mark in that section to clarify it. This is substantiated by the fact that this is not optional in a similar example later in the tutorial, see https://github.com/Shopify/graphql-design-tutorial/blob/master/TUTORIAL.md#manipulating-relationships